requirements/requirements-local.txt
asgiref==3.10.0
Django==5.2.8
sqlparse==0.5.3
psycopg2-binary==2.9.9
Pillow==10.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.12.0

